+++ tests/ml_perf_end_to_end_tests.py
@@ -8,9 +8,9 @@ import subprocess
 import numpy as np
 import pandas as pd
 
-from mplc import multi_partner_learning
 from mplc.corruption import Duplication
 from mplc.experiment import Experiment
+from mplc.multi_partner_learning import basic_mpl
 from mplc.scenario import Scenario
 from . import test_utils
 
@@ -38,7 +38,7 @@ class Test_EndToEndTest:
         titanic_scenario_1 = Scenario(2, [0.4, 0.6], epoch_count=3, minibatch_count=1, dataset='titanic')
         titanic_scenario_2 = Scenario(3, [0.2, 0.2, 0.6],
                                       corruption_parameters=['not-corrupted', corruption_1, 'not-corrupted'],
-                                      epoch_count=3, minibatch_count=1, dataset_name='titanic')
+                                      epoch_count=3, minibatch_count=1, dataset='titanic')
         exp = Experiment(experiment_name='end_to_end_titanic',
                          scenarios_list=[titanic_scenario_1],
                          nb_repeats=2,
@@ -52,11 +52,11 @@ class Test_EndToEndTest:
 
     def test_all_mpl_approaches(self):
         """
-        Test all the mpl approaches
+        Test all the multi_partner_learning approaches
         """
 
         exp = Experiment()
-        mpl_approaches = multi_partner_learning.MULTI_PARTNER_LEARNING_APPROACHES.copy()
+        mpl_approaches = basic_mpl.MULTI_PARTNER_LEARNING_APPROACHES.copy()
 
         for approach in mpl_approaches:
             exp.add_scenario(Scenario(2, [0.25, 0.75], epoch_count=2, minibatch_count=2, dataset='mnist',
